src/test/java/com/qa/services/CustomerServiceTest.java
package com.qa.services;
import org.junit.Test;
import org.junit.runner.RunWith;
import org.mockito.InjectMocks;
import org.mockito.Mock;
import org.mockito.Mockito;
import org.mockito.junit.MockitoJUnitRunner;


import com.qa.persistence.Dao;
import domain.Customer;
@	RunWith(MockitoJUnitRunner.class)
public class CustomerServiceTest {

		@Mock
		private Dao<Customer> customerDao;
		
		@InjectMocks
		private CustomerService customerServices;
		
		@Test
		public void customerServicesCreate() {
			Customer customer = new Customer("chris", "perrins");
			customerServices.create(customer);
			Mockito.verify(customerDao, Mockito.times(1)).create(customer);
		}
		
		@Test
		public void customerServicesRead() {
			customerServices.readAll();
			Mockito.verify(customerDao, Mockito.times(1)).readAll();
		}
		
		@Test
		public void customerServicesUpdate() {
			Customer customer = new Customer("chris", "perrins");
			customerServices.update(customer);
			Mockito.verify(customerDao, Mockito.times(1)).update(customer);
		}
		
		@Test
		public void customerServicesDelete() {
			customerServices.delete(1L);;
			Mockito.verify(customerDao, Mockito.times(1)).delete(1L);
		}
	}
